Move compilation from make_app to start_server to compile apps and well as components
