+++ src/test/make_app_test.ts
@@ -57,77 +57,4 @@ suite('makeApp', () => {
     assert.equal(app.packageName, 'no_bower_json');
   });
 
-  suite('compilation', () => {
-
-    const testCompilation = (options: {
-      url: string,
-      agent?: string,
-      compile: 'always' | 'never' | 'auto',
-      result: 'compiled' | 'uncompiled'
-    }) => async() => {
-      const url = options.url;
-      const agent = options.agent;
-      const compile = options.compile;
-      const result = options.result;
-      const app = makeApp({
-        root: root,
-        componentDir: path.join(root, 'bower_components'), compile,
-      });
-      let request = supertest(app).get(url);
-      if (agent) {
-        request = request.set('User-Agent', agent);
-      }
-      const response = await request;
-      const isCompiled = response.text.indexOf('class A {}') === -1;
-      const shouldCompile = result === 'compiled';
-      if (isCompiled && !shouldCompile) {
-        throw new Error('Source was compiled');
-      } else if (!isCompiled && shouldCompile) {
-        throw new Error('Source was not compiled');
-      }
-    };
-
-    test('compiles external JS when --compile=always', testCompilation({
-           url: '/test-component/test.js',
-           compile: 'always',
-           result: 'compiled',
-         }));
-
-    test('compiles inline JS when --compile=always', testCompilation({
-           url: '/test-component/test.html',
-           compile: 'always',
-           result: 'compiled',
-         }));
-
-    test('doesn\'t compile external JS when --compile=never', testCompilation({
-           url: '/test-component/test.js',
-           compile: 'never',
-           result: 'uncompiled',
-         }));
-
-    test('doesn\'t compile inline JS when --compile=never', testCompilation({
-           url: '/test-component/test.html',
-           compile: 'never',
-           result: 'uncompiled',
-         }));
-
-    test(
-        'doesn\'t compile external JS when --compile=auto and agent is Chrome',
-        testCompilation({
-          url: '/test-component/test.js',
-          agent:
-              'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_12_1) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/55.0.2883.52 Safari/537.36',
-          compile: 'auto',
-          result: 'uncompiled',
-        }));
-
-    test(
-        'compiles external JS when --compile=auto and agent is unknown',
-        testCompilation({
-          url: '/test-component/test.js',
-          compile: 'auto',
-          result: 'compiled',
-        }));
-  });
-
 });
